renderização consertada, renderização parcial de objetos fora da tela

+++ internal/game/entities/entity.go
@@ -3,11 +3,11 @@ package entities
 import "github.com/hajimehoshi/ebiten/v2"
 
 type Entity struct {
-	X, Y   float64
+	X, Y   int
 	Sprite *ebiten.Image
 }
 
-func (e *Entity) move(dx, dy float64) {
+func (e *Entity) move(dx, dy int) {
 	e.X += dx
 	e.Y += dy
 }
